De-archangel: move EmailTemplate to judgels.mailer
Move the last remaining file in judgels.jophiel (main source) to its
natural home next to Mailer/MailerConfiguration/MailerModule. With this
the judgels-server-app/src/main/java/judgels/jophiel/ directory is gone
and the judgels.jophiel main-source package is fully eliminated.

- judgels.jophiel.EmailTemplate -> judgels.mailer.EmailTemplate
- File moved via git mv (history preserved). Package declaration
  rewritten. Immutables-generated ImmutableEmailTemplate regenerates
  under the new package automatically.
- 4 import sites updated:
  - judgels.user.account.UserResetPasswordConfiguration
  - judgels.user.account.UserResetPasswordMailer
  - judgels.contrib.user.registration.UserRegistrationConfiguration
  - judgels.contrib.user.registration.UserRegistrationEmailMailer

Verified: checkstyle (main/test/integTest), compileJava,
compileTestJava, compileIntegTestJava, and :judgels-server-app:test
all pass.
