+++ judgels-backends/judgels-server-app/src/main/java/judgels/mailer/EmailTemplate.java
@@ -1,4 +1,4 @@
-package judgels.jophiel;
+package judgels.mailer;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.immutables.value.Value;
