+++ judgels-backends/judgels-server-app/src/main/java/judgels/user/account/UserResetPasswordConfiguration.java
@@ -1,7 +1,7 @@
 package judgels.user.account;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import judgels.jophiel.EmailTemplate;
+import judgels.mailer.EmailTemplate;
 import org.immutables.value.Value;
 
 @Value.Immutable
