+++ judgels-backends/judgels-server-app/src/main/java/judgels/user/account/UserResetPasswordMailer.java
@@ -1,7 +1,7 @@
 package judgels.user.account;
 
 import judgels.api.user.User;
-import judgels.jophiel.EmailTemplate;
+import judgels.mailer.EmailTemplate;
 import judgels.mailer.Mailer;
 
 public class UserResetPasswordMailer {
